requirements.txt
neo4j==5.17.0
openai==1.23.0
requests==2.31.0
ollama==0.1.8
python-dotenv==1.0.1